add retirement time type and start to fix menu
Lots of tweaks to make the code work with updated menu system, which apparently just changes the API for the fun of it without documenting in the change log.
Also add in beginnings of RetirementTime data type to handle updating memory location & displaying time-to-retirement

Menu is still mostly broken, but code to wipe and reinit EEPROM works, and a completely unnecessary fade-out fake initializing percent counter added.